Modify data type

+++ src/EE/data.py
@@ -5,7 +5,7 @@ class GASTScore(Data):
     def __init__(self, practitioners):
         self._practitioners = practitioners
 
-    def getIndividualScore(self, person: str) -> int:
+    def getIndividualScore(self, person: str) -> float:
         return self._practitioners[common_name][1]
 
 
@@ -13,7 +13,7 @@ class SchoolScore(Data):
     def __init__(self, practitioners):
         self._practitioners = practitioners
 
-    def getIndividualScore(self, person: str) -> int:
+    def getIndividualScore(self, person: str) -> float:
 
         return self._practitioners[common_name][1]
 
@@ -22,7 +22,7 @@ class ProgramType(Data):
     def __init__(self, practitioners):
         self._practitioners = practitioners
 
-    def getIndividualScore(self, person: str) -> int:
+    def getIndividualScore(self, person: str) -> float:
         return self._practitioners[common_name][1]
 
 
@@ -30,7 +30,7 @@ class Competition(Data):
     def __init__(self, practitioners):
         self._practitioners = practitioners
 
-    def getIndividualScore(self, person: str) -> int:
+    def getIndividualScore(self, person: str) -> float:
         return self._practitioners[common_name][1]
 
 
@@ -38,7 +38,7 @@ class SchoolActivity(Data):
     def __init__(self, practitioners):
         self._practitioners = practitioners
 
-    def getIndividualScore(self, person: str) -> int:
+    def getIndividualScore(self, person: str) -> float:
         return self._practitioners[common_name][1]
